pyproject.toml
[project]
name = "mode-connectivity"
version = "0.1.0"
description = ""
authors = [
    {name = "Maciej Lodzinski",email = "mf.lodzinski@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "torch",
    "torchvision",
    "hydra-core",
    "wandb",
    "numpy",
    "scipy",
    "matplotlib",
    "rich",
    "tabulate",
    "torchaudio (>=2.9.1,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "umap-learn (>=0.5.9.post2,<0.6.0)",
    "imageio (>=2.37.2,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
